requirements.txt
# ============================================================================
# VOXENT v2.0 Requirements - Consolidated and Deduplicated
# Python 3.8+ required
# ============================================================================

# ============================================================================
# CORE AUDIO PROCESSING
# ============================================================================
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
pyaudio>=0.2.13
audioread>=3.0.0

# ============================================================================
# DEEP LEARNING & GPU SUPPORT
# ============================================================================
# NOTE: Install PyTorch via conda for Windows compatibility:
# conda install pytorch torchvision torchaudio cuda-toolkit -c pytorch -c nvidia -y
torch>=2.1.0
torchaudio>=2.1.0
torchvision>=0.15.0

# ============================================================================
# SPEAKER DIARIZATION
# ============================================================================
pyannote.audio>=3.1.0

# ============================================================================
# SPEECH & ML PROCESSING
# ============================================================================
speechbrain>=0.5.13
scikit-learn>=1.3.0

# ============================================================================
# SOURCE SEPARATION (Optional)
# ============================================================================
demucs>=4.0.0

# ============================================================================
# DATA & ML UTILITIES
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
joblib>=1.3.0

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
flask>=3.0.0
werkzeug>=3.0.0
flask-socketio>=5.3.0

# ============================================================================
# CONFIGURATION & UTILITIES
# ============================================================================
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.0
pathlib>=1.0.1

# ============================================================================
# OPTIONAL: DATA ANALYSIS & VISUALIZATION
# ============================================================================
matplotlib>=3.7.0     # For visualization
plotly>=5.17.0        # Interactive plots

# ============================================================================
# OPTIONAL: GPU/INFERENCE OPTIMIZATION
# ============================================================================
onnx>=1.14.0          # Model optimization
onnxruntime>=1.16.0   # ONNX runtime
